Move NavbarItemLists into their own folders
