+++ front/app/modules/commercial/customizable_navbar/admin/containers/VisibleNavbarItemList/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 // services
-import { reorderNavbarItem, removeNavbarItem } from '../../services/navbar';
+import { reorderNavbarItem, removeNavbarItem } from '../../../services/navbar';
 import { deletePage } from 'services/pages';
 import { INavbarItem, getNavbarItemSlug } from 'services/navbar';
 
@@ -12,7 +12,7 @@ import {
   LockedRow,
 } from 'components/admin/ResourceList';
 import { SubSectionTitle } from 'components/admin/Section';
-import NavbarItemRow from '../components/NavbarItemRow';
+import NavbarItemRow from '../../components/NavbarItemRow';
 
 // hooks
 import useNavbarItems from 'hooks/useNavbarItems';
@@ -22,7 +22,7 @@ import usePageSlugById from 'hooks/usePageSlugById';
 // i18n
 import { injectIntl, FormattedMessage } from 'utils/cl-intl';
 import { InjectedIntlProps } from 'react-intl';
-import messages from './messages';
+import messages from '../messages';
 
 // utils
 import { isNilOrError } from 'utils/helperUtils';
